Consolidate benchmarks into a single test_benchmark.py
